% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFa.R
\name{corFA}
\alias{corFA}
\title{Insert Communalities in the Diagonal of a Correlation or a Covariance Matrix}
\usage{
corFA(R, method = "ginv")
}
\arguments{
\item{R}{numeric: correlation matrix.}

\item{method}{character: actually only \code{"ginv"} is supplied to compute the approximation of the communalities by maximum correlation.}
}
\value{
\item{values}{ numeric: matrix of correlation/covariance with communalities in the diagonal. }
}
\description{
This function inserts communalities in the diagonal of a correlation/covariance matrix.
}
\examples{
\dontshow{set.seed(123); x   <- matrix(rnorm(1000),ncol=10);}
\donttest{corFA(x, method = "ginv")}

}
\seealso{
\code{\link{nScree}}
}
