% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenFrom.R
\name{eigenFrom}
\alias{eigenFrom}
\title{Identify the Data Type to Obtain the Eigenvalues}
\usage{
eigenFrom(x)
}
\arguments{
\item{x}{numeric: a \code{vector} of eigenvalues, a \code{matrix} of correlations or of covariances or a \code{data.frame} of data}
}
\value{
\item{value}{ character: return the data type to obtain the eigenvalues: \code{"eigenvalues"}, \code{"correlation"} or \code{"data"} }
}
\description{
The \code{eigenFrom} function identifies the data type from which to obtain the eigenvalues. The function is used internally in many functions of the package to be able to apply these to a vector of eigenvalues, a matrix of correlations or covariance or a \code{data.frame}.
}
\examples{
\dontshow{set.seed(123); x   <- matrix(rnorm(1000),ncol=10);}
\donttest{eigenFrom(x)}
}
