% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_many_factors.R
\name{how_many_factors}
\alias{how_many_factors}
\title{Plots a Scree Plot to determine how many factors (or componentes to retain)}
\usage{
how_many_factors(df, rep = 1000, cent = 0.05)
}
\arguments{
\item{df}{a dataframe with only \code{int} or \code{num} type of variables}

\item{rep}{numeric: number of replications of the correlation matrix (default is 1000)}

\item{cent}{numeric: quantile of the distribution on which the decision is made (default is 0.05)}
}
\description{
\code{how_many_factors()} call upon the \code{\link[nFactors]{plotnScree}} function from \code{nFactors} package to iterate over the variables of a dataframe.
}
\details{
Determine how many factors to retain based on:
\enumerate{
\item{Eigenvalues}
\item{Parallel Analysis}
\item{Optimal Coordinates}
\item{Acceleration Factor}
}
}
\seealso{
\link[nFactors]{plotnScree}
}
