% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreStats.R
\name{moreStats}
\alias{moreStats}
\title{Statistical Summary of a Data Frame}
\usage{
moreStats(x, quantile = 0.95, show = FALSE)
}
\arguments{
\item{x}{numeric: matrix or \code{data.frame}.}

\item{quantile}{numeric: quantile of the distribution.}

\item{show}{logical: if \code{TRUE} prints the quantile choosen.}
}
\value{
\item{values}{ numeric: \code{data.frame} of statistics: mean, median, quantile, standard deviation, minimum and maximum. }
}
\description{
This function produces another summary of a \code{data.frame}. This function was proposed in order to apply some functions globally on a \code{data.frame}: \code{quantile}, \code{median}, \code{min} and \code{max}. The usual \emph{R} version cannot do so.
}
\examples{
\dontshow{set.seed(123); x   <- matrix(rnorm(1000),ncol=10);}
\donttest{results <- moreStats(x, quantile=0.05, show=TRUE)}

}
\seealso{
\code{\link{nScree}}
}
