% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel}
\alias{parallel}
\title{Parallel Analysis of a Correlation or Covariance Matrix}
\usage{
parallel(subject = 100, var = 10, rep = 1000, quantile = 0.05,
  model = "components", sd = diag(1, var), ...)
}
\arguments{
\item{subject}{numeric: nmber of subjects (default is 100)}

\item{var}{numeric: number of variables (default is 10)}

\item{rep}{numeric: number of replications of the correlation matrix (default is 1000)}

\item{quantile}{numeric: quantile of the distribution on which the decision is made (default is 0.05)}

\item{model}{character: \code{"components"} or \code{"factors"}}

\item{sd}{numeric: vector of standard deviations of the simulated variables
(for a parallel analysis on a covariance matrix)}

\item{...}{variable: other parameters for the \code{"mvrnorm"}, \code{corr} or \code{cov} functions}
}
\value{
A list with \enumerate{
\item{eigen}{ Data frame consisting of the mean and the quantile of
the eigenvalues distribution }
\item{eigen$mevpea}{ Mean of the eigenvalues distribution}
\item{eigen$sevpea}{ Standard deviation of the eigenvalues distribution}
\item{eigen$qevpea}{ quantile of the eigenvalues distribution}
\item{eigen$sqevpea}{ Standard error of the quantile of the eigenvalues distribution}
\item{subject}{ Number of subjects}
\item{variables}{ Number of variables}
\item{quantile}{ Selected quantile}
}
}
\description{
This function gives the distribution of the eigenvalues of correlation or a
covariance matrices of random uncorrelated standardized normal variables.
The mean and a selected quantile of this distribution are returned.
}
\details{
Note that if the decision is based on a quantile value rather than on the
mean, care must be taken with the number of replications (\code{rep}). In
fact, the smaller the quantile (\code{quantile}), the bigger the number of
necessary replications.
}
\examples{
\dontshow{set.seed(123); df <- as.data.frame(matrix(rnorm(1000),100,10));}
\donttest{
ev <- eigen(cor(df)) # get eigenvalues
eig <- ev$values # eigenvalues
ap <- parallel(subject = nrow(df), var = ncol(df), rep = 1000, quantile = .05, model = "components")
}
}
\references{
Drasgow, F. and Lissak, R. (1983) Modified parallel analysis: a
procedure for examining the latent dimensionality of dichotomously scored
item responses. \emph{Journal of Applied Psychology, 68}(3), 363-373.

Hoyle, R. H. and Duvall, J. L. (2004). Determining the number of factors in
exploratory and confirmatory factor analysis.  In D. Kaplan (Ed.): \emph{The
Sage handbook of quantitative methodology for the social sciences}. Thousand
Oaks, CA: Sage.

Horn, J. L. (1965). A rationale and test of the number of factors in factor
analysis. \emph{Psychometrika, 30}, 179-185.
}
