% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLoadings.R
\name{printLoadings}
\alias{printLoadings}
\title{Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff}
\usage{
printLoadings(x, digits = 3, cutoff = 0.4, sort = TRUE, ...)
}
\arguments{
\item{x}{object resultant from a factor analysis \code{\link[psych]{principal}} or \code{\link[psych]{fa}} function from \code{psych} package}

\item{digits}{number of digits to print}

\item{cutoff}{minimal variable loadings to show}

\item{sort}{binary True or False}

\item{...}{etc}
}
\description{
Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff
}
\examples{
\dontrun{printLoadings(fa.fit, digits = 3, cutoff = 0.4, sort = TRUE)}
}
\seealso{
taken from: \href{https://stackoverflow.com/questions/12114440/how-do-i-export-a-sorted-factor-loading-table}{StackOverflow}
}
