% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screePlotAPA.R
\name{screePlotAPA}
\alias{screePlotAPA}
\title{APA Scree Plot for Parallel Analysis}
\usage{
screePlotAPA(x, rep = 1000, quantile = 0.05, model = "components")
}
\arguments{
\item{x}{a matrix or dataframe}

\item{rep}{numeric: number of replications of the correlation matrix (default is 1000)}

\item{quantile}{numeric: quantile of the distribution on which the decision is made (default is 0.05)}

\item{model}{character: \code{"components"} or \code{"factors"}}
}
\value{
APA-friendly plot of Parallel Analysis
}
\description{
\code{screePlotAPA()} plots a friendly APA Scree Plot with the measured eigenvalues and the quantile of the eigenvalues distribution from a parallel analysis
}
\examples{
\dontshow{set.seed(123); df <- as.data.frame(matrix(rnorm(1000),100,10));}\donttest{screePlotAPA(df)}
}
