% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenComputes.R
\name{eigenComputes}
\alias{eigenComputes}
\title{Computes Eigenvalues According to the Data Type}
\usage{
eigenComputes(x, cor = TRUE, model = "components", ...)
}
\arguments{
\item{x}{numeric: a \code{vector} of eigenvalues, a \code{matrix} of correlations or of covariances or a \code{data.frame} of data}

\item{cor}{logical: if \code{TRUE} computes eigenvalues from a correlation matrix, else from a covariance matrix}

\item{model}{character: \code{"components"} or \code{"factors"}}

\item{...}{variable: additionnal parameters to give to the \code{cor} or \code{cov} functions}
}
\value{
\item{value}{ numeric: return a vector of eigenvalues }
}
\description{
The \code{eigenComputes} function computes eigenvalues from the identified data type. It is used internally in many functions o in order to apply these to a vector of eigenvalues, a matrix of correlations or covariance or a data frame.
}
\examples{
\dontshow{set.seed(123); x   <- matrix(rnorm(1000),ncol=10);}
\donttest{
eigenComputes(x, cor=TRUE,  use="everything")
eigenComputes(x, cor=FALSE, use="everything")
eigenComputes(x, cor=TRUE,  use="everything", method="spearman")
eigenComputes(x, cor=TRUE,  use="everything", method="kendall")
}
}
