% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLoadings.R
\name{printLoadings}
\alias{printLoadings}
\title{Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff}
\usage{
printLoadings(x, digits = 3, cutoff = 0.4, sort = TRUE, ...)
}
\description{
Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff
}
\seealso{
taken from: https://stackoverflow.com/questions/12114440/how-do-i-export-a-sorted-factor-loading-table
}
