% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLoadings.R
\name{printLoadings}
\alias{printLoadings}
\title{Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff}
\usage{
printLoadings(x, digits = 3, cutoff = 0.4, sort = TRUE, ...)
}
\arguments{
\item{x}{object resultant from a factor analysis \code{\link[psych]{principal}} or \code{\link[psych]{fa}} function from \code{psych} package}

\item{digits}{number of digits to print}

\item{cutoff}{minimal variable loadings to show}

\item{sort}{binary True or False}

\item{...}{etc}
}
\value{
A matrix or dataframe that can be exported as a CSV file with \code{\link[utils]{write.table}}
}
\description{
Helper function for printing rotated component matrix
Function to get a rotated component matrix from a factor analysis in psych while sorting and cutoff
}
\examples{
set.seed(123)
df <- as.data.frame(matrix(rnorm(100*10, 1, .5), ncol=10))
fit <- principal(df, nfactors = 2)
printLoadings(fit$loadings, digits = 3, cutoff = 0.4, sort = TRUE)

}
